+++ backend/products/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField()
+    price = models.BigIntegerField()
     multiple = models.IntegerField(blank=True, null=True)
     is_offer = models.BooleanField(default=False)
     image_url = models.URLField(blank=True, null=True)
